requirements.txt
torch~=2.2.2
openai~=1.54.4
llama-index-core~=0.12.0
pathlib~=1.0.1
pydantic~=2.9.2
chainlit~=1.3.2
python-dotenv~=1.0.1
minio~=7.2.10
aiofiles~=23.2.1
aiohttp~=3.11.2
SQLAlchemy~=2.0.36
pillow~=11.0.0
python-pptx~=1.0.2
fastapi~=0.115.5
pymilvus~=2.5.4
starlette~=0.41.2
httpx~=0.27.2
flask
vanna[snowflake]
vanna[postgres]
db-dtypes
python-dotenv